+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added Site-to-site example
 - Removed `compression` from examples
 - Switched to bash wizards
+- Added auth-pass-verify example script
 
 ### 2.0.6 - Fixed bugs, added additonal parameters